content/python-oauth-databricks-app/requirements.txt
databricks-sql-connector==3.0.1
databricks-sdk==0.20.0
numpy<2.0.0
posit-sdk==0.3.1
shiny==0.7.1
